Update Cache flush, and try to use Dave2D display list properly
Rather than just rendering each Dave2D operation as it is called,
try to create a display list with as many operation in it as possible,
before executing the display list. If SW drawing is used, then the
display list has to be executed before the SW drawing is execute to
preserve the drawing order.